Nick/branch42 (#233)
* fix invalid country for localized pricing point

* archiving and deletion system for posts

* revamped moderator queue logic

* using workflows for all post functions

* fix workflow structure to be more streamlined

* fix some more stuff, make resource ids nullable by default

* added unarchiving posts, generate resource urls concurrently, and fixed up post moderator queue

* fixing up tests with new workflows in parley

* fix up tests with new formats

* update schema, fix issues

* fix and update schema types, use ints to store currency values instead of floats

* fix broken tests

* add email to principal, fix test not working

* add check eslint, fix some ts issues

* fix dataconverter test

* add a link, dont run ccbill tests for now

* remove email for now since its weird

* add build file

* fix seeder

+++ applications/hades/internal/domain/billing/price.go
@@ -2,18 +2,18 @@ package billing
 
 type Price struct {
 	currency Currency
-	amount   float64
+	amount   int64
 }
 
 func (p *Price) Currency() Currency {
 	return p.currency
 }
 
-func (p *Price) Amount() float64 {
+func (p *Price) Amount() int64 {
 	return p.amount
 }
 
-func UnmarshalPricingFromDatabase(currency Currency, amount float64) *Price {
+func UnmarshalPricingFromDatabase(currency Currency, amount int64) *Price {
 	return &Price{
 		currency: currency,
 		amount:   amount,
